broke out properties and form into their own components, banner is now dynamic
