+++ app/models/portfolio_edit.rb
@@ -1,0 +1,2 @@
+class PortfolioEdit < ApplicationRecord
+end
