+++ db/migrate/20200525175551_create_portfolio_edits.rb
@@ -1,0 +1,11 @@
+class CreatePortfolioEdits < ActiveRecord::Migration[5.2]
+  def change
+    create_table :portfolio_edits do |t|
+      t.string :bannerImage
+      t.string :headerText1
+      t.string :headerText2
+
+      t.timestamps
+    end
+  end
+end
