+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2020_05_22_185907) do
+ActiveRecord::Schema.define(version: 2020_05_25_175551) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -139,6 +139,14 @@ ActiveRecord::Schema.define(version: 2020_05_22_185907) do
     t.string "line1"
     t.string "line2"
     t.string "line3"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+  end
+
+  create_table "portfolio_edits", force: :cascade do |t|
+    t.string "bannerImage"
+    t.string "headerText1"
+    t.string "headerText2"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
