+++ test/models/portfolio_edit_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class PortfolioEditTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
